+++ src/main/java/ssrahoo/marketplaceapi/dto/UserUpdateDto.java
@@ -1,7 +1,9 @@
 package ssrahoo.marketplaceapi.dto;
 
+import java.math.BigDecimal;
+
 public record UserUpdateDto(
         String username,
         String password,
-        Double wallet
+        BigDecimal wallet
 ) {}
